sort existing test of helpers
